// beeclient/src/types.ts
export interface Job {
  id: number;
  job_title: string;
  company_name: string;
  company_rating: string;
  company_reviews: string;
  locations: string;
  job_salary: string;
  experience: string;
  full_description: string;
  role: string;
  department: string;
  employment_type: string;
  role_category: string;
  education_ug: string;
  education_pg: string;
  key_skills: string[];
  apply_redirect_url: string;
  naukri_url: string;
  posted_on_listing: string;
  created_at: string;
}